Collection Log Luck - Refactor luck settings to work for all drops
